+++ src/birdnet_tests/models/model_v2m4_protobuf_py/test_predict_species_within_audio_file_gpu.py
@@ -1,12 +1,14 @@
 
+from typing import Optional
+
 import pytest
 import tensorflow as tf
 
 from birdnet.models.model_v2m4_protobuf import ModelV2M4Protobuf
 from birdnet_tests.models.test_predict_species_within_audio_file import (
+  model_minimum_test_soundscape_predictions_are_correct,
   model_test_identical_predictions_return_same_result,
-  model_test_soundscape_predictions_are_correct,
-  model_test_soundscape_predictions_batch_size_4_are_correct)
+  model_test_soundscape_predictions_are_correct)
 
 
 @pytest.fixture(name="model")
@@ -19,13 +21,16 @@ def get_model():
   return None
 
 
-def test_soundscape_predictions_are_correct(model: ModelV2M4Protobuf):
-  model_test_soundscape_predictions_are_correct(model)
+def test_soundscape_predictions_are_correct(model: Optional[ModelV2M4Protobuf]):
+  if model is not None:
+    model_test_soundscape_predictions_are_correct(model, precision=2)
 
 
-def test_soundscape_predictions_batch_size_4_are_correct(model: ModelV2M4Protobuf):
-  model_test_soundscape_predictions_batch_size_4_are_correct(model)
+def test_minimum_test_soundscape_predictions_are_correct(model: Optional[ModelV2M4Protobuf]):
+  if model is not None:
+    model_minimum_test_soundscape_predictions_are_correct(model, precision=2)
 
 
-def test_identical_predictions_return_same_result(model: ModelV2M4Protobuf):
-  model_test_identical_predictions_return_same_result(model)
+def test_identical_predictions_return_same_result(model: Optional[ModelV2M4Protobuf]):
+  if model is not None:
+    model_test_identical_predictions_return_same_result(model)