+++ src/birdnet_tests/models/model_v2m4_tflite_py/test_predict_species_within_audio_file.py
@@ -5,9 +5,9 @@ import pytest
 from birdnet.models.model_v2m4_tflite import ModelV2M4TFLite
 from birdnet.types import Species
 from birdnet_tests.models.test_predict_species_within_audio_file import (
-  TEST_FILE_WAV, model_test_identical_predictions_return_same_result,
-  model_test_soundscape_predictions_are_correct,
-  model_test_soundscape_predictions_batch_size_4_are_correct)
+  TEST_FILE_WAV, model_minimum_test_soundscape_predictions_are_correct,
+  model_test_identical_predictions_return_same_result,
+  model_test_soundscape_predictions_are_correct)
 
 
 @pytest.fixture(name="model")
@@ -17,11 +17,11 @@ def get_model():
 
 
 def test_soundscape_predictions_are_correct(model: ModelV2M4TFLite):
-  model_test_soundscape_predictions_are_correct(model)
+  model_test_soundscape_predictions_are_correct(model, precision=2)
 
 
-def test_soundscape_predictions_batch_size_4_are_correct(model: ModelV2M4TFLite):
-  model_test_soundscape_predictions_batch_size_4_are_correct(model)
+def test_minimum_test_soundscape_predictions_are_correct(model: ModelV2M4TFLite):
+  model_minimum_test_soundscape_predictions_are_correct(model, precision=2)
 
 
 def test_identical_predictions_return_same_result(model: ModelV2M4TFLite):